Change login code to use common request code.
This produces safer JSON for the request. It does introduce the
possibility of an infinite loop if the UniFi controller returns the
"login required" signal to a login attempt, but that seems unlikely.
If it does happen, we can solve it with another reqOpts field to
disable the automatic login logic.